+++ src/DiagnosticMessages.ts
@@ -145,7 +145,7 @@ export let DiagnosticMessages = {
     }),
     statementMustBeDeclaredAtTopOfFile: (statementKeyword: string) => ({
         message: `'${statementKeyword}' statement must be declared at the top of the file`,
-        code: 1149,
+        code: 1024,
         severity: DiagnosticSeverity.Error
     }),
     methodDoesNotExistOnType: (methodName: string, className: string) => ({
@@ -730,9 +730,21 @@ export let DiagnosticMessages = {
         code: 1139,
         severity: DiagnosticSeverity.Error
     }),
-    itemCannotBeUsedAsType: (typeText: string) => ({
-        message: `'${typeText}' cannot be used as a type`,
+    /**
+     *
+     * @param name for function calls where we can't find the name of the function
+     * @param fullName if a namespaced name, this is the full name `alpha.beta.charlie`, otherwise it's the same as `name`
+     * @param typeName if 'name' refers to a member, what is the the type it is a member of?
+     * @param typeDescriptor defaults to 'type' ... could also be 'namespace', etc.
+     */
+    cannotFindFunction: (name: string, fullName?: string, typeName?: string, typeDescriptor = 'type') => ({
+        message: `Cannot find function '${name}'${typeName ? ` for ${typeDescriptor} '${typeName}'` : ''}`,
         code: 1140,
+        data: {
+            name: name,
+            fullName: fullName ?? name,
+            typeName: typeName ? typeName : undefined
+        },
         severity: DiagnosticSeverity.Error
     }),
     argumentTypeMismatch: (actualTypeString: string, expectedTypeString: string, data?: TypeCompatibilityData) => ({
@@ -778,19 +790,24 @@ export let DiagnosticMessages = {
         code: 1148,
         severity: DiagnosticSeverity.Error
     }),
+    itemCannotBeUsedAsType: (typeText: string) => ({
+        message: `'${typeText}' cannot be used as a type`,
+        code: 1149,
+        severity: DiagnosticSeverity.Error
+    }),
     expectedNewlineInConditionalCompile: () => ({
         message: `Expected newline in conditional compilation statement`,
-        code: 1150,
+        code: 1151,
         severity: DiagnosticSeverity.Error
     }),
     expectedTerminatorOnConditionalCompileBlock: () => ({
         message: `Expected '#end if', '#else if', or '#else' to terminate conditional compilation block`,
-        code: 1151,
+        code: 1152,
         severity: DiagnosticSeverity.Error
     }),
     unsafeUnmatchedTerminatorInConditionalCompileBlock: (terminator: string) => ({
         message: `Unsafe unmatched terminator '${terminator}' in conditional compilation block`,
-        code: 1152,
+        code: 1153,
         severity: DiagnosticSeverity.Error
     })
 };